requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.5
pydantic>=2.5.0
httpx>=0.25.2
python-multipart>=0.0.6

